+omni shadow set check
To render omni shadows, the omni shadow shader must be set. Rendering omni shadows is optional for point lights and spotlights in case omni shader is active. If the omni shadow shader is not set, the omni shadow pass will not work. In that case, if point and spot lights have shadow maps, unexpected results may occur, so their shadowmaps should not be set.